xen: Implement hypercall for tracing of program counters
This commit makes the changes to the hypervisor, the build system as
well as libxc necessary in order to facilitate tracing of program counters.

A discussion of the design can be found in the mailing list:
https://lists.xen.org/archives/html/xen-devel/2017-05/threads.html#02210

The list of files to be included for tracing might still be too extensive,
resulting in indeterministic tracing output for some use cases.

Signed-off-by: Felix Schmoll <eggi.innovations@gmail.com>

---
Changed since v1:
 * Fixed bug that prevented xen from compiling when CONFIG_TRACE_PC was disabled
 * Adapted formatting to coding style
 * Renamed files and variables to conform to trace_pc-idiom
 * Changed hypercall variables to be publicly accessible
 * Added appropriate licenses
 * Fix position of hypercall in hypercall-table
 * Increase tracing coverage
 * Delete buffer if on domain destruction path